Small refactor for pdf dialogs to make it probably more robust.
