+++ resources/assets/js/pdfDialog.js
@@ -1,22 +1,42 @@
-window.initPdfModalScripts = function () {
+class PdfDialog {
 
-    const options = () => {
-        return $.param({
-            'headings': $('#pdfHeadings').prop('checked'),
-            'nums': $('#pdfNums').prop('checked'),
-            'refs': $('#pdfRefs').prop('checked'),
-            'quantity': $('#pdfQuantity').val()
+    init() {
+
+        const options = () => {
+            return $.param({
+                'headings': $('#pdfHeadings').prop('checked'),
+                'nums': $('#pdfNums').prop('checked'),
+                'refs': $('#pdfRefs').prop('checked'),
+                'quantity': $('#pdfQuantity').val()
+            });
+        };
+
+        $('#pdfModal').on('shown.bs.modal', (event) => {
+            const button = event.relatedTarget;
+            const recipient = button.getAttribute('data-bs-view');
+            fetch(`${recipient}`)
+                .then(response => response.text())
+                .then(data => {
+                    const modalContent = pdfModal.querySelector('.modal-body');
+                    modalContent.innerHTML = `${data}`;
+                    $("#pdfDownload").off('click');
+                    $("#pdfDownload").on('click', (event) => {
+                        const ref = $('#previewContainer').data('ref');
+                        const translationId = $('#previewContainer').data('translation');            
+                        window.open(`/pdf/ref/${translationId}/${ref}?${options()}`);
+                        $('#pdfDownload').blur();
+                        bootstrap.Modal.getInstance($('#pdfModal')).hide();
+                    });        
+                }
+                )   .catch((e) => {
+                    console.log("Error loading content", e);
+                }
+                );
         });
     };
 
-    $('#pdfModal').on('shown.bs.modal', (event) => {
-        const ref = $('#previewContainer').data('ref');
-        const translationId = $('#previewContainer').data('translation');
-        $("#pdfDownload").off('click');        
-        $("#pdfDownload").on('click', (event) => {
-            window.open(`/pdf/ref/${translationId}/${ref}?${options()}`);
-            $('#pdfDownload').blur();
-            bootstrap.Modal.getInstance($('#pdfModal')).hide();
-        });
-    });
-};
+}
+
+export default function initPdfModal() {
+    new PdfDialog().init();
+}
